break up the reporters
